refactor: Namespace the API client
