+++ src/api/rest/providers/github/endpoints/index.js
@@ -1,0 +1,7 @@
+import * as orgs from './orgs';
+import * as search from './search';
+
+export default {
+  orgs,
+  search,
+};
